♻️ refactor type
