+++ src/types/gitHub.ts
@@ -1,0 +1,10 @@
+export type GitHubCardProps = {
+  id: string;
+  name: string;
+  description: string;
+  forkCount: number;
+  language: string;
+  pinnedItems?: any;
+  stargazerCount: number;
+  url: string;
+};
