+++ src/types/twitter.ts
@@ -1,0 +1,12 @@
+export type TwitterUserProps = {
+  id: string;
+  name: string;
+  profile_image_url: string;
+  username: string;
+};
+
+export type TwitterTweetProps = {
+  id: string;
+  created_at: string;
+  text: string;
+};
